waviwiet win

+++ src/pgdp/adventuin/AdventuinParty.java
@@ -135,7 +135,7 @@ public final class AdventuinParty {
                   positiveQuantity++;
               }
               double resultat = 0.0;
-              if (negativeQuantity != 0) resultat += (-1.0) * (double) negativeSum / negativeQuantity;
+              if (negativeQuantity != 0) resultat += (-(double)1.0) * (double) negativeSum / negativeQuantity;
               if (positiveQuantity != 0) resultat += (double) positiveSum / positiveQuantity;
